+++ R/ternary_contour.R
@@ -13,9 +13,6 @@ setMethod(
                         ilr = TRUE, method = "linear", extrapolate = FALSE,
                         palette = function(i) grDevices::hcl.colors(i, "YlOrRd", rev = TRUE),
                         ...) {
-    ## /!\ For backward compatibility with akima::interp() /!\
-    if (isTRUE(list(...)$linear)) method <- "linear"
-
     ## Calculate contour lines
     xy <- coordinates_contour(x = x, y = y, z = z, value = value, n = n,
                               nlevels = nlevels, levels = levels,